requirements.txt
huggingface_hub==0.26.5
numpy==2.2.0
openai==1.57.2
pandas==2.2.3
python-dotenv==1.0.1
